# documentation/api-docs.yaml
openapi: 3.0.0
info:
  title: CRM Rules Engine API
  description: API documentation for the CRM Rules Engine microservices
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  schemas:
...
    Log:
      type: object
      properties:
        id:
          type: string
        rule_id:
          type: string
        user_id:
          type: string
        action:
          type: string
        status:
          type: string
          enum: [success, failure]
        executed_at:
          type: string
          format: date-time
    GatewayResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          enum: ["success", "failed"]
          example: "success"
        data:
          type: object
...
paths:
...
  /logs:
    get:
      tags:
        - Logs
      summary: Get execution logs
      parameters:
        - name: rule_id
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of execution logs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GatewayResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Log'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GatewayResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 400
                      message:
                        type: string
                        enum: ["failed"]
                      data:
                        type: object
                        properties:
                          error:
                            type: string
                            example: "invalid request"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GatewayResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 500
                      message:
                        type: string
                        enum: ["success"]
                      data:
                        type: object
                        properties:
                          error:
                            type: string
                            example: "internal server error"
  /logs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Logs
      summary: Get a log entry by ID
      responses:
        '200':
          description: Log entry details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GatewayResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Log'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GatewayResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 404
                      message:
                        type: string
                        enum: ["failed"]
                      data:
                        type: object
                        properties:
                          error:
                            type: string
                            example: "log not found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/GatewayResponse'
                  - type: object
                    properties:
                      code:
                        type: integer
                        example: 500
                      message:
                        type: string
                        enum: ["success"]
                      data:
                        type: object
                        properties:
                          error:
                            type: string
                            example: "internal server error"